feat: add localStorage hook
